Fix spurious creation of package srcdir

+++ rock/git_hook.rb
@@ -1,5 +1,9 @@
 module Rock
     def self.install_git_hook(pkg, source_file, target_hook)
+        if !File.directory?(pkg.srcdir)
+            raise ArgumentError, "cannot install the git hooks for a non-checked out package"
+        end
+
         hook_source_path = File.join(
             File.expand_path(File.dirname(__FILE__)),
             source_file)
@@ -8,6 +12,9 @@ module Rock
         FileUtils.cp hook_source_path, hook_dest_path
     end
     def self.remove_git_hook(pkg, target_hook)
+        if !File.directory?(pkg.srcdir)
+            raise ArgumentError, "cannot remove the git hooks for a non-checked out package"
+        end
         hook_dest_path = File.join(pkg.srcdir, '.git', 'hooks', target_hook)
         FileUtils.rm_f hook_dest_path
     end
